// frontend/src/app/model/Intermediaire.ts
export class Intermediaire {
    
    inter_numero: number ;
    inter_type: number ;
    inter_denomination: String ;
	inter_denominationcourt: String ;
	inter_boitepostale: String ;
	inter_rue: String ;
	inter_quartierville: String ;
	inter_telephone1: String ;
	inter_telephone2: String ;
	inter_portable: String ;
	inter_email: String ;
	inter_classificationmetier: String ;
	inter_codecommission: number ;
	inter_objectifcaannuel: number ;
	inter_caportefeuille: number ;
	inter_sinistralite: number ;
	inter_arriere: number ;
	inter_dureemoyenne: number ;
	inter_montantcommission: number ;
	inter_codeutilisateur: String ;
	inter_datemodification: Date ;
	inter_numagrement: String;
	inter_datentrerelation: Date;
	inter_datedebutcarteagrement:Date;
	inter_autorisation : string;
	inter_anneeexercice:number;
}